+++ src/start.py
@@ -16,7 +16,7 @@ class Launcher(object):
         Create a new Launcher. PYTHONPATH is set and current work directory has changed.
         :return: a new Launcher's instance
         """
-        self._app_path = os.sep.join(os.path.abspath(__file__).split(os.sep)[:-3])
+        self._app_path = os.sep.join(os.path.abspath(__file__).split(os.sep)[:-2])
         self._app = None
         self._main_widget = None
 
@@ -50,10 +50,10 @@ class Launcher(object):
         Launch the application.
         :return: exit code of the application (int)
         """
-        import src.core.gui
+        import src.core.app
 
         self._app = Qtwidgets.QApplication(sys.argv)
-        self._main_widget = src.core.gui.TwentyThreeTools()
+        self._main_widget = src.core.app.TwentyThreeTools()
 
         self._main_widget.show()
         return self._app.exec_()
